A bunch of creature features, calibration of tip location works, i'm happy with it's performance

+++ CircleDetector.cs
@@ -11,14 +11,14 @@ namespace Picky
     {
         public double Param1;
         public double Param2;
-        public double Exposure;
+        public int Exposure;
         public HoughModes DetectorType;
 
         public double zEstimate;                    //In mm to optical plane
         public CircleSegment CircleEstimate;        //In mm
         public OpenCvSharp.Rect ROI;                //In mm
 
-        public CircleDetector(HoughModes mode, double param1, double param2, double exposure)
+        public CircleDetector(HoughModes mode, double param1, double param2, int exposure)
         {
             DetectorType = mode;
             Param1 = param1;
